feat: fixed comments
