+++ src/pages/add-material-page/types.ts
@@ -1,0 +1,5 @@
+export interface IMaterialConfig {
+  title: string;
+  path: string;
+  buttonText: string;
+}
