--- src/shared/attribute-manager/index.ts
@@ -1,1 +1,0 @@
-export * from './attribute-manager';
