+++ src/shared/button/button.tsx
@@ -1,11 +1,8 @@
-import {
-  Button as BaseButton,
-  type ButtonProps as MUIButtonProps,
-} from '@mui/material';
+import { type ButtonProps, Button as BaseButton } from '@mui/material';
 
 import { useStyles } from './styles';
 
-interface IButtonProps extends Omit<MUIButtonProps, 'variant'> {
+interface IButtonProps extends Omit<ButtonProps, 'variant'> {
   variant?: 'primary' | 'secondary';
 }
