+++ src/shared/dynamic-values-table/index.ts
@@ -1,0 +1,1 @@
+export * from './dynamic-values-table';
